// src/app/models/swadesh-categories.ts
export enum Category {
    conjunction = 'conjunction',
    adjective = 'adjective',
    noun = 'noun',
    verb = 'verb'
}

export enum Subcategory {
    personal_pronoun = 'personal pronoun',
    demonstrative_pronoun = 'demonstrative pronoun',
    interrogative_pronoun = 'interrogative pronoun',
    adverb = 'adverb',
    preposition = 'preposition',
    quantity = 'quantity',
    number = 'number',
    color = 'color',
    size = 'size',
    state = 'state',
    family = 'family',
    animal = 'animal',
    nature = 'nature',
    body = 'body',
    weather = 'weather',
    landscape = 'landscape',
    mouth = 'mouth',
    animation = 'animation',
    life = 'life',
    survival = 'survival',
    action = 'action',
    mobility = 'mobility',
    other = 'other'
}