// src/app/models/swadesh-item.ts
export interface SwadeshItem {
    category: string;
    subcategory: string;
    english: string;
    dutch: string;
    german: string;
    danish: string;
    french: string;
    italian: string;
    spanish: string;
    portuguese: string;
    russian: string;
    japanese: string;
    greek: string;
}